configs: apple: Add custom bootcmd to boot from preferred ESP
AsahiLinux uses a distinct EFI system partition per OS installation 1).
To boot the intended system u-boot has to prefer the intended partition
for booting. The previous stage bootloader specifies in UUID of the
partition in the device tree as "/chosen/asahi,efi-system-partition".
'apple_bootcmd' reads the UUID from the device tree and uses modified
distro boot scripts to only boot from a partition with a matching UUID.
If it fails it tries to boot from the first USB device.

1: https://github.com/AsahiLinux/docs/wiki/Open-OS-Ecosystem-on-Apple-Silicon-Macs

Signed-off-by: Janne Grunau <j@jannau.net>